+++ src/games/wow/wdc3file.cpp
@@ -121,8 +121,8 @@ bool WDC3File::open()
 #endif
   }
 
-  uint32 palletBlockOffset = getPos();
-  uint32 commonBlockOffset = palletBlockOffset + m_header.pallet_data_size;
+  size_t palletBlockOffset = getPos();
+  size_t commonBlockOffset = palletBlockOffset + m_header.pallet_data_size;
  
   // this only supports one section. We need to be able to loop and support multiple sections.
   seek(sectionHeader[0].file_offset);
@@ -130,7 +130,7 @@ bool WDC3File::open()
   data = getPointer();
 
   // compute various offset needed to read data in the file 
-  uint32 stringTableOffset = getPos() + recordSize * recordCount;
+  size_t stringTableOffset = getPos() + recordSize * recordCount;
 
   // embedded strings in fields instead of stringTable
   if ((m_header.flags & 0x01) != 0)
@@ -142,11 +142,11 @@ bool WDC3File::open()
   seek(stringTableOffset);
   stringTable = getPointer();
 
-  uint32 IdBlockOffset = stringTableOffset + stringSize;
+  size_t IdBlockOffset = stringTableOffset + stringSize;
 
-  uint32 copyBlockOffset = IdBlockOffset + sectionHeader[0].id_list_size;
+  size_t copyBlockOffset = IdBlockOffset + sectionHeader[0].id_list_size;
 
-  uint32 relationshipDataOffset = copyBlockOffset + sectionHeader[0].copy_table_count * 8;
+  size_t relationshipDataOffset = copyBlockOffset + sectionHeader[0].copy_table_count * 8;
  
 #if WDC3_READ_DEBUG > 2
   LOG_INFO << "m_header.flags & 0x01" << (m_header.flags & 0x01);
@@ -678,7 +678,7 @@ bool WDC3File::readFieldValue(size_t recordIndex, unsigned int fieldIndex, uint 
     {                                          
       uint32 index = readBitpackedValue(info, recordOffset);
       auto it = m_palletBlockOffsets.find(fieldIndex);
-      uint32 offset = it->second + index * 4;
+      size_t offset = it->second + (size_t)index * 4;
       memcpy(&result, getBuffer() + offset, 4);
       break;
     }
@@ -686,7 +686,7 @@ bool WDC3File::readFieldValue(size_t recordIndex, unsigned int fieldIndex, uint 
     {
       uint32 index = readBitpackedValue(info, recordOffset);
       auto it = m_palletBlockOffsets.find(fieldIndex);
-      uint32 offset = it->second + index * arraySize * 4 + arrayIndex * 4;
+      size_t offset = it->second + (size_t)index * arraySize * 4 + arrayIndex * 4;
       memcpy(&result, getBuffer() + offset, 4);
       break;
     }
